# Add APIs that remove redundant segments from paths