--- a/src/libraries/System.Private.CoreLib/src/System/IO/RedundantSegmentHelper.Windows.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/RedundantSegmentHelper.Windows.cs
@@ -34,8 +34,11 @@ namespace System.IO
 
             int prefixLength = isDevicePath ? PathInternal.DevicePrefixLength : 0;
             ReadOnlySpan<char> pathWithoutPrefix = originalPath.Slice(prefixLength);
-            int rootLength = PathInternal.GetRootLength(pathWithoutPrefix);
-            int prefixAndRootLength = prefixLength + rootLength;
+
+            // In a path like "\\.\C:..\folder\subfolder\file.txt":
+            // - PathInternal.GetRootLength(originalPath) will return "\\.\C:..\", which includes the extra redundant segment.
+            // - PathInternal.GetRootLength(pathWithoutPrefix) will return "C:", which will help to determine later if a path was rooted or not.
+            int prefixAndRootLength = prefixLength + PathInternal.GetRootLength(pathWithoutPrefix);
 
             // Append characters that should not be touched:
             // - "C:" - known drive, unknown root