--- a/src/libraries/System.Private.CoreLib/src/System/IO/RedundantSegmentHelper.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/RedundantSegmentHelper.cs
@@ -1,6 +1,5 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
 
 #nullable enable
 using System.Diagnostics;