--- a/src/libraries/System.Private.CoreLib/src/System/IO/RedundantSegmentHelper.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/IO/RedundantSegmentHelper.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 #nullable enable
-using System.Diagnostics;
 using System.Text;
 
 namespace System.IO