… src/libraries/Common/tests/Tests/System/IO/PathInternal.Tests.cs
… src/libraries/System.Private.CoreLib/src/System/IO/PathInternal.cs
… src/libraries/System.Runtime.Extensions/tests/System.Runtime.Extensions.Tests.csproj
… src/libraries/System.Runtime.Extensions/tests/System/IO/PathTests_Windows.cs
… src/libraries/System.Runtime.Extensions/tests/System/IO/RedundantSegmentsTestsBase.cs
… src/libraries/System.Runtime.Extensions/tests/System/IO/RedundantSegmentsTests_Unix.cs
… src/libraries/System.Runtime.Extensions/tests/System/IO/RedundantSegmentsTests_Windows.cs
… src/libraries/System.Runtime/ref/System.Runtime.cs