Fixes: https://github.com/dotnet/runtime/issues/2162

The APIs added in this PR are:

```csharp
namespace System.IO
{
    public class Path
    {
        public static string RemoveRedundantSegments(string path);
        public static string RemoveRedundantSegments(ReadOnlySpan<char> path);
        public static bool TryRemoveRedundantSegments(ReadOnlySpan<char> path, Span<char> destination, out int charsWritten);
    }
}
```

The original internal methods were not able to handle all the edge cases.

The unit tests for both Windows and Unix are passing, as well as all the previously existing unit tests for `Path.GetFullPath`, which is the main method in `System.IO` that was consuming the formerly internal method that removed redundant segments.

The code is different between Linux and Windows because there are many special cases in Windows.

Unix rules:
- `/` at the beginning of the path means the path is rooted. It's also the segment separator. Repeated contiguous separators get merged into one.
- "." refers the current directory. Gets removed except when it's the first character in a path without a known root.
- ".." backtracks to the previous directory. The previous segment needs to be removed, and the current ".." segment does not get added. The exception is when the path root `/` is unknown and either there are no more directories to backtrack, or all other previous segments are also "..", in which case, the current segment must also stay.
- `\` is a valid file or folder character. For example: `one\segment` is a single file or folder name.
- A segment whose name consists of 3 or more dots is a valid segment name.

Windows rules:
- Both `\` and `/` are considered path separators. The `/` character gets automatically normalized to `\` when detected. Repeated contiguous separators get merged into one.
- "." refers to the current directory. Gets removed except when it's the first character in a path without a known root.
- ".." backtracks to the previous directory. The previous segment needs to be removed, and the current ".." segment does not get added. The exception is when the path root is unknown and either there are no more directories to backtrack, or all other previous segments are also "..", in which case, the current segment must also stay.
- A segment whose name consist of 3 or more dots is a valid segment name, unless it's the last segment in the path and is not followed by a trailing separator, in which case the segment gets removed.
- A segment whose name has trailing dots gets the trailing dots removed.
- Paths that begin with a device prefix (`\\.\`, `\\?\` or `\??\`) have some special rules:
    - A segment whose name consist of 3 or more dots never get removed.
    - A segment whose name has trailing dots do not get the dots removed.
- Paths that have a drive but not a separator are considered unqualified. Example: "C:folder".
- Paths that start with a separator but have no drive are considered rooted. Example: `\folder`. 
- `PathInternal.GetRootLength` already contains the logic to decide what portion of the path should be considered the root. It can vary depending if the path begins with a device prefix, or is in the `\\Server\Share` format, or if the root has no drive, or if the root has no separator.

Windows paths are explained in detail in these sources:

- File path formats on Windows systems - https://docs.microsoft.com/en-us/dotnet/standard/io/file-path-formats
- From @JeremyKuhne 's blog:
    - Path Format Overview - https://docs.microsoft.com/en-us/archive/blogs/jeremykuhne/path-format-overview
    - Path Normalization - https://docs.microsoft.com/en-us/archive/blogs/jeremykuhne/path-normalization
    - DOS to NT - https://docs.microsoft.com/en-us/archive/blogs/jeremykuhne/dos-to-nt-a-paths-journey