@carlossanlop 2020-06-17
I'm investigating why the net472 leg is failing.