@carlossanlop 2020-07-11
@jkotas @Jozkee @jeffhandley here is the perf PR: https://github.com/dotnet/performance/pull/1394


### Results

#### master branch

|                              Method |                                 path |     Mean |   Error |  StdDev |   Median |      Min |      Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------ |------------------------------------- |---------:|--------:|--------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
| GetFullPathWithoutRedundantSegments |  ..\.\..\.\.\(...)..\.\.\..\\. [104] | 527.9 ns | 6.47 ns | 5.73 ns | 528.5 ns | 518.3 ns | 537.8 ns | 0.0471 |     - |     - |     304 B |
| GetFullPathWithoutRedundantSegments | ..\.\..\.\.\f(...)3\.\....\..\. [92] | 518.4 ns | 8.76 ns | 7.77 ns | 516.0 ns | 505.8 ns | 532.7 ns | 0.0473 |     - |     - |     304 B |
| GetFullPathWithoutRedundantSegments | C:\..\.\.\fol(...)3\.\....\..\. [90] | 243.5 ns | 2.14 ns | 2.00 ns | 243.4 ns | 240.6 ns | 246.5 ns | 0.0127 |     - |     - |      80 B |
| GetFullPathWithoutRedundantSegments |  \\.\Server\S(...)..\.\.\..\\. [109] | 291.5 ns | 4.25 ns | 3.98 ns | 291.3 ns | 286.3 ns | 299.0 ns | 0.0162 |     - |     - |     104 B |


#### RRS branch (this PR)

|                              Method |                                 path |     Mean |   Error |  StdDev |   Median |      Min |      Max |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------ |------------------------------------- |---------:|--------:|--------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
| GetFullPathWithoutRedundantSegments |  ..\.\..\.\.\(...)..\.\.\..\\. [104] | 516.9 ns | 3.70 ns | 3.09 ns | 516.8 ns | 512.0 ns | 523.9 ns | 0.0478 |     - |     - |     304 B |
| GetFullPathWithoutRedundantSegments | ..\.\..\.\.\f(...)3\.\....\..\. [92] | 507.8 ns | 7.01 ns | 6.56 ns | 506.7 ns | 499.1 ns | 521.7 ns | 0.0467 |     - |     - |     304 B |
| GetFullPathWithoutRedundantSegments | C:\..\.\.\fol(...)3\.\....\..\. [90] | 240.8 ns | 1.81 ns | 1.51 ns | 240.5 ns | 238.4 ns | 243.8 ns | 0.0126 |     - |     - |      80 B |
| GetFullPathWithoutRedundantSegments |  \\.\Server\S(...)..\.\.\..\\. [109] | 285.8 ns | 1.69 ns | 1.58 ns | 286.0 ns | 283.5 ns | 288.4 ns | 0.0159 |     - |     - |     104 B |


#### Comparison
❯ dotnet run --base "D:\perf_before\" --diff "D:\perf_after\" --threshold 0.01%

summary:
better: 4, geomean: 1.018
total diff: 4

No Slower results for the provided threshold = 0.01% and noise filter = 0.3ns.

| Faster                                                                               | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| System.IO.Tests.Perf_Path.GetFullPathWithoutRedundantSegments(path: "..\\.\\..\\.\\. |      1.02 |           528.53 |           516.76 |         |
| System.IO.Tests.Perf_Path.GetFullPathWithoutRedundantSegments(path: "\\\\.\\Server\\ |      1.02 |           291.31 |           286.01 |         |
| System.IO.Tests.Perf_Path.GetFullPathWithoutRedundantSegments(path: "..\\.\\..\\.\\. |      1.02 |           516.01 |           506.69 |         |
| System.IO.Tests.Perf_Path.GetFullPathWithoutRedundantSegments(path: "C:\\..\\.\\.\\f |      1.01 |           243.39 |           240.52 |         |