@jkotas 2020-07-11
You performance test is Windows specific and it is testing atypical case that is very unlikely to appear in the real world and . Corner cases like this one are good for functional testing, but they are not appropriate for performance testing.

You should focus performance testing on common real world use cases of this API, and also cover both Windows and Unix.

E.g. I have done a quick check on Linux that run `Path.GetFullPath("/home/jkotas/runtime/src/coreclr/runtime/src/libraries/System.Private.CoreLib/src/System/IO/Path.cs");`. I see 1.5x regression on this one with your change.