@carlossanlop 2020-11-26
I checked-out this branch in my WSL with Ubuntu. I ran all the Path benchmarks, where I also made sure to include a benchmark testing the path shared by @jkotas above ([see PR](https://github.com/dotnet/performance/pull/1394)), and the new benchmark did not show up:

```
carlos@calopepc:~/performance/src/tools/ResultsComparer$ dotnet run -c release --base "/home/carlos/perf_before/" --diff "/home/carlos/perf_after/" --threshold 0.01%
summary:
better: 4, geomean: 1.077
worse: 3, geomean: 1.053
total diff: 7

| Slower                                     | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| System.IO.Tests.Perf_Path.GetTempPath      |      1.12 |           157.33 |           175.82 |         |
| System.IO.Tests.Perf_Path.GetDirectoryName |      1.02 |            33.14 |            33.91 |         |
| System.IO.Tests.Perf_Path.GetFileName      |      1.02 |            31.73 |            32.39 |         |

| Faster                                                 | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| System.IO.Tests.Perf_Path.Combine                      |      1.28 |             3.84 |             3.00 |         |
| System.IO.Tests.Perf_Path.ChangeExtension              |      1.04 |            21.01 |            20.26 |         |
| System.IO.Tests.Perf_Path.GetFullPathForReallyLongPath |      1.01 |          3067.42 |          3039.39 |         |
| System.IO.Tests.Perf_Path.GetFullPathForLegacyLength   |      1.00 |           610.23 |           608.12 |         |
```