@jkotas 2020-11-26
I still see the regression for the example I gave above. Can you run this program before and after changes:
```
using System.IO;
using System.Diagnostics;

var sw = new Stopwatch();
sw.Start();
for (int i = 0; i < 100000000; i++)
    Path.GetFullPath("/home/jkotas/runtime/src/libraries/System.Private.CoreLib/src/System/IO/Path.cs");
sw.Stop();
Console.WriteLine(sw.ElapsedMilliseconds.ToString());    
```
I prints `41390` with your changes and `23331` without your changes in my WSL2 Ubuntu instance.