@carlossanlop 2021-01-19
@jkotas I modified `Path.GetFullPath` so that it is now calling the `RemoveRedundantSegments` overload that takes a string, instead of a span. That way, when there is no redundancy, the original string is returned.

I ran your code snippet in my Mac, and there was only a slight improvement:

- Without my changes: 24886 ms
- With my changes: 44872 ms
- With my changes, but calling the string overload: 40983 ms

Would it be possible to get the changes merged as they are right now so we can get potential feedback in Preview 1? I can open an issue to make additional performance improvements later.