@jkotas 2021-01-19
> Would it be possible to get the changes merged as they are right now so we can get potential feedback in Preview 1?

Who do you expect to use these APIs and to get a feedback from?