@carlossanlop 2021-01-19
The [original proposal](https://github.com/dotnet/runtime/issues/2162) had some links from people interested in it: https://github.com/dotnet/runtime/issues/15584, https://github.com/MonoGame/MonoGame/issues/7167. There was also an issue from the arcade repo I think, but right now I can't find it.